backend/requirements.txt
sanic
sanic-ext
tortoise-orm
aiomysql
cryptography
aioredis
bcrypt
pyjwt
python-dotenv
requests
pytest